Simplify handling updates of table settings across components
