+++ frontend/src/panels/dreammall/TableSetupStepB.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="flat-text-field d-flex flex-column align-center pa-4 w-100">
     <v-text-field
-      v-model="tableNameModel"
+      v-model="tableSettings.name"
       rounded
       class="elevation-0 w-100"
       content-class="elevation-0"
@@ -12,35 +12,22 @@
     />
     <!-- todo: manage values as maxlength globally? -->
 
-    <v-switch v-model="isPrivateModel" label="Privat" color="#4caf50" inset hide-details />
+    <v-switch v-model="tableSettings.isPrivate" label="Privat" color="#4caf50" inset hide-details />
 
     <SimpleButton class="mt-12" :label="submitText" @click="onNext" />
   </div>
 </template>
 
 <script setup lang="ts">
-import { computed } from 'vue'
-
 import SimpleButton from '#components/buttons/SimpleButton.vue'
 
-import { TableSetupEmits, TableSetupProps } from './TableSetupProps'
+import type MyTableSettings from '#src/panels/dreammall/interfaces/MyTableSettings'
+import type { TableSetupProps, TableSetupEmits } from './TableSetupProps'
 
-const props = defineProps<TableSetupProps>()
+defineProps<TableSetupProps>()
 const emit = defineEmits<TableSetupEmits>()
 
-const tableNameModel = computed({
-  get: () => props.myTableSettings.name,
-  set: (value) => {
-    emit('tableName:updated', value)
-  },
-})
-
-const isPrivateModel = computed({
-  get: () => !props.myTableSettings.isPublic,
-  set: (value) => {
-    emit('isPublic:updated', !value)
-  },
-})
+const tableSettings = defineModel<MyTableSettings>({ required: true })
 
 const onNext = () => {
   emit('next')
